fix(chat-composer): address review — disabled-nav, naming, quote reveal
Syncs the feat-applicable subset of the #16260 review fixes (@0xfullex). The
paste-preference and i18n items are carve-only (feat's composer uses the
pref-gated services/PasteService and already has the keys); the AgentComposer
failed-send skill-cache fix needs a restore branch feat doesn't have yet.

- QuickPanel keyboard nav (components/QuickPanel/view.tsx) routes Arrow/Page
  through moveQuickPanelSelectableIndex so the highlight skips disabled rows
  (Enter no longer silently no-ops); deleted the dead parallel nav family from
  list.tsx (QuickPanelList/Frame/toggleQuickPanelSelectedId/CandidateItem) +
  the Tab/Enter collapse guard now respects manageListExternally.
- Naming: …API -> …Api in ComposerToolProvider; extracted
  findLatestPendingAskUserQuestionRequest into askUserQuestionComposerRequest.ts
  so the test base matches its subject.
- Quote insertion reveals the composer (toggleExpanded(true)) instead of
  re-asserting the current expand state (a no-op).
