+++ src/renderer/components/QuickPanel/__tests__/list.test.ts
@@ -1,25 +1,18 @@
 import { describe, expect, it } from 'vitest'
 
-import {
-  firstQuickPanelSelectableIndex,
-  moveQuickPanelSelectableIndex,
-  type QuickPanelCandidateItem,
-  toggleQuickPanelSelectedId
-} from '../list'
+import { firstQuickPanelSelectableIndex, moveQuickPanelSelectableIndex } from '../list'
 
-const createItems = (): QuickPanelCandidateItem[] => [
-  { id: 'one', label: 'One' },
-  { id: 'disabled', label: 'Disabled', disabled: true },
-  { id: 'two', label: 'Two' },
-  { id: 'three', label: 'Three' },
-  { id: 'four', label: 'Four' }
+const createItems = (): { id: string; disabled?: boolean }[] => [
+  { id: 'one' },
+  { id: 'disabled', disabled: true },
+  { id: 'two' },
+  { id: 'three' },
+  { id: 'four' }
 ]
 
 describe('QuickPanel list primitives', () => {
   it('finds the first selectable item', () => {
-    expect(
-      firstQuickPanelSelectableIndex([{ id: 'disabled', label: 'Disabled', disabled: true }, ...createItems()])
-    ).toBe(1)
+    expect(firstQuickPanelSelectableIndex([{ id: 'disabled', disabled: true }, ...createItems()])).toBe(1)
   })
 
   it('moves by one with wrapping while skipping disabled items', () => {
@@ -35,15 +28,5 @@ describe('QuickPanel list primitives', () => {
     expect(moveQuickPanelSelectableIndex(items, 0, 2, { wrap: false })).toBe(3)
     expect(moveQuickPanelSelectableIndex(items, 3, 2, { wrap: false })).toBe(4)
     expect(moveQuickPanelSelectableIndex(items, 3, -2, { wrap: false })).toBe(0)
-  })
-
-  it('toggles selected ids immutably', () => {
-    const selectedIds = new Set(['one'])
-    const withoutOne = toggleQuickPanelSelectedId(selectedIds, 'one')
-    const withTwo = toggleQuickPanelSelectedId(withoutOne, 'two')
-
-    expect([...selectedIds]).toEqual(['one'])
-    expect(withoutOne.has('one')).toBe(false)
-    expect(withTwo.has('two')).toBe(true)
   })
 })
