+++ src/renderer/components/chat/composer/variants/shared/composerQuote.ts
@@ -22,18 +22,16 @@ interface QuoteInsertionActions {
 /**
  * Subscribes to the main-process quote IPC and inserts the quoted text as a quote token via
  * the composer's imperative actions ref. The insertion runs through `useEffectEvent` so the
- * IPC listener subscribes once and never re-subscribes when `isExpanded` toggles.
+ * IPC listener subscribes once and stays stable across renders.
  */
-export function useComposerQuoteInsertion<T extends QuoteInsertionActions>(
-  actionsRef: RefObject<T>,
-  isExpanded: boolean
-): void {
+export function useComposerQuoteInsertion<T extends QuoteInsertionActions>(actionsRef: RefObject<T>): void {
   const { t } = useTranslation()
 
   const insertQuote = useEffectEvent((selectedText: string) => {
     if (!selectedText) return
     actionsRef.current.insertToken(createQuoteToken(selectedText, t('selection.action.builtin.quote')))
-    actionsRef.current.toggleExpanded(isExpanded)
+    // Reveal the composer so the freshly inserted quote is visible even when collapsed.
+    actionsRef.current.toggleExpanded(true)
   })
 
   useEffect(() => {
